[PP-1274] Removes ODL Hold Reaper and replaces with ODL 2 Hold Reaper. Also the dashboard active holds query has been modified to hide any expired holds.
