+++ docker/services/cron/cron.d/circulation
@@ -81,11 +81,11 @@ HOME=/var/www/circulation
 # OPDS 1.x + ODL
 #
 15 * * * * root bin/run odl_import_monitor >> /var/log/cron.log 2>&1
-0 */8 * * * root bin/run odl_hold_reaper >> /var/log/cron.log 2>&1
 
 # OPDS 2.x + ODL import
 #
 45 * * * * root bin/run odl2_import_monitor >> /var/log/cron.log 2>&1
+0 */8 * * * root bin/run odl2_hold_reaper >> /var/log/cron.log 2>&1
 
 # SAML
 #
