+++ src/palace/manager/api/odl.py
@@ -45,8 +45,7 @@ from palace.manager.core.lcp.credential import (
     LCPHashedPassphrase,
     LCPUnhashedPassphrase,
 )
-from palace.manager.core.metadata_layer import FormatData, LicenseData, TimestampData
-from palace.manager.core.monitor import CollectionMonitor
+from palace.manager.core.metadata_layer import FormatData, LicenseData
 from palace.manager.core.opds_import import (
     BaseOPDSImporter,
     OPDSImporter,
@@ -527,6 +526,7 @@ class BaseODLAPI(PatronActivityCirculationAPI[SettingsType, LibrarySettingsType]
         # We have successfully borrowed this book.
         if hold:
             _db.delete(hold)
+            # log circulation event:  hold converted to loan
         self.update_licensepool(licensepool)
         return loan
 
@@ -866,6 +866,8 @@ class BaseODLAPI(PatronActivityCirculationAPI[SettingsType, LibrarySettingsType]
         _db = Session.object_session(hold)
         licensepool = hold.license_pool
         _db.delete(hold)
+
+        # log a circulation event : hold_released
         self.update_licensepool(licensepool)
         return True
 
@@ -899,6 +901,7 @@ class BaseODLAPI(PatronActivityCirculationAPI[SettingsType, LibrarySettingsType]
         for hold in holds:
             if hold.end and hold.end < utc_now():
                 _db.delete(hold)
+                # log circulation event:  hold expired
                 self.update_licensepool(hold.license_pool)
             else:
                 self._update_hold_data(hold)
@@ -1327,48 +1330,3 @@ class ODLImportMonitor(OPDSImportMonitor):
         super().__init__(
             _db, collection, import_class, force_reimport=True, **import_class_kwargs
         )
-
-
-class ODLHoldReaper(CollectionMonitor):
-    """Check for holds that have expired and delete them, and update
-    the holds queues for their pools."""
-
-    SERVICE_NAME = "ODL Hold Reaper"
-    PROTOCOL = ODLAPI.label()
-
-    def __init__(
-        self,
-        _db: Session,
-        collection: Collection,
-        api: ODLAPI | None = None,
-        **kwargs: Any,
-    ):
-        super().__init__(_db, collection, **kwargs)
-        self.api = api or ODLAPI(_db, collection)
-
-    def run_once(self, progress: TimestampData) -> TimestampData:
-        # Find holds that have expired.
-        expired_holds = (
-            self._db.query(Hold)
-            .join(Hold.license_pool)
-            .filter(LicensePool.collection_id == self.api.collection_id)
-            .filter(Hold.end < utc_now())
-            .filter(Hold.position == 0)
-        )
-
-        changed_pools = set()
-        total_deleted_holds = 0
-        for hold in expired_holds:
-            changed_pools.add(hold.license_pool)
-            self._db.delete(hold)
-            total_deleted_holds += 1
-
-        for pool in changed_pools:
-            self.api.update_licensepool(pool)
-
-        message = "Holds deleted: %d. License pools updated: %d" % (
-            total_deleted_holds,
-            len(changed_pools),
-        )
-        progress = TimestampData(achievements=message)
-        return progress
