+++ src/palace/manager/api/odl2.py
@@ -20,7 +20,8 @@ from palace.manager.api.odl import (
     ODLLibrarySettings,
     ODLSettings,
 )
-from palace.manager.core.metadata_layer import FormatData
+from palace.manager.core.metadata_layer import FormatData, TimestampData
+from palace.manager.core.monitor import CollectionMonitor
 from palace.manager.core.opds2_import import (
     OPDS2Importer,
     OPDS2ImporterSettings,
@@ -35,16 +36,17 @@ from palace.manager.integration.settings import (
 from palace.manager.sqlalchemy.model.collection import Collection
 from palace.manager.sqlalchemy.model.edition import Edition
 from palace.manager.sqlalchemy.model.licensing import LicensePool, RightsStatus
+from palace.manager.sqlalchemy.model.patron import Hold
 from palace.manager.sqlalchemy.model.resource import HttpResponseTuple
 from palace.manager.util import first_or_default
-from palace.manager.util.datetime_helpers import to_utc
+from palace.manager.util.datetime_helpers import to_utc, utc_now
 
 if TYPE_CHECKING:
     from webpub_manifest_parser.core.ast import Metadata
     from webpub_manifest_parser.opds2.ast import OPDS2Feed, OPDS2Publication
 
     from palace.manager.api.circulation import HoldInfo
-    from palace.manager.sqlalchemy.model.patron import Hold, Loan, Patron
+    from palace.manager.sqlalchemy.model.patron import Loan, Patron
 
 
 class ODL2Settings(ODLSettings, OPDS2ImporterSettings):
@@ -324,3 +326,49 @@ class ODL2ImportMonitor(OPDS2ImportMonitor):
         super().__init__(
             _db, collection, import_class, force_reimport=True, **import_class_kwargs
         )
+
+
+class ODL2HoldReaper(CollectionMonitor):
+    """Check for holds that have expired and delete them, and update
+    the holds queues for their pools."""
+
+    SERVICE_NAME = "ODL2 Hold Reaper"
+    PROTOCOL = ODL2API.label()
+
+    def __init__(
+        self,
+        _db: Session,
+        collection: Collection,
+        api: ODL2API | None = None,
+        **kwargs: Any,
+    ):
+        super().__init__(_db, collection, **kwargs)
+        self.api = api or ODL2API(_db, collection)
+
+    def run_once(self, progress: TimestampData) -> TimestampData:
+        # Find holds that have expired.
+        expired_holds = (
+            self._db.query(Hold)
+            .join(Hold.license_pool)
+            .filter(LicensePool.collection_id == self.api.collection_id)
+            .filter(Hold.end < utc_now())
+            .filter(Hold.position == 0)
+        )
+
+        changed_pools = set()
+        total_deleted_holds = 0
+        for hold in expired_holds:
+            changed_pools.add(hold.license_pool)
+            self._db.delete(hold)
+            # log circulation event:  hold expired
+            total_deleted_holds += 1
+
+        for pool in changed_pools:
+            self.api.update_licensepool(pool)
+
+        message = "Holds deleted: %d. License pools updated: %d" % (
+            total_deleted_holds,
+            len(changed_pools),
+        )
+        progress = TimestampData(achievements=message)
+        return progress
