+++ tests/manager/api/test_odl2.py
@@ -12,14 +12,16 @@ from palace.manager.api.circulation_exceptions import (
     PatronHoldLimitReached,
     PatronLoanLimitReached,
 )
-from palace.manager.api.odl2 import ODL2Importer
+from palace.manager.api.odl2 import ODL2HoldReaper, ODL2Importer
 from palace.manager.core.coverage import CoverageFailure
 from palace.manager.sqlalchemy.constants import (
     EditionConstants,
     IdentifierConstants,
     MediaTypes,
 )
+from palace.manager.sqlalchemy.model.collection import Collection
 from palace.manager.sqlalchemy.model.contributor import Contribution, Contributor
+from palace.manager.sqlalchemy.model.datasource import DataSource
 from palace.manager.sqlalchemy.model.edition import Edition
 from palace.manager.sqlalchemy.model.licensing import (
     DeliveryMechanism,
@@ -30,6 +32,7 @@ from palace.manager.sqlalchemy.model.patron import Hold
 from palace.manager.sqlalchemy.model.resource import Hyperlink
 from palace.manager.sqlalchemy.model.work import Work
 from palace.manager.sqlalchemy.util import create
+from palace.manager.util.datetime_helpers import utc_now
 from tests.fixtures.api_odl import (
     LicenseHelper,
     LicenseInfoHelper,
@@ -37,7 +40,7 @@ from tests.fixtures.api_odl import (
     ODL2APIFilesFixture,
 )
 from tests.fixtures.database import DatabaseTransactionFixture
-from tests.fixtures.odl import ODL2APITestFixture
+from tests.fixtures.odl import ODL2APITestFixture, ODL2TestFixture
 
 
 class TestODL2Importer:
@@ -437,3 +440,56 @@ class TestODL2API:
         with pytest.raises(PatronHoldLimitReached) as exc:
             odl2api.api.place_hold(odl2api.patron, "pin", pool, "")
         assert exc.value.limit == 1
+
+
+class TestODL2HoldReaper:
+    def test_run_once(
+        self, odl2_test_fixture: ODL2TestFixture, db: DatabaseTransactionFixture
+    ):
+        library = odl2_test_fixture.library()
+        collection = odl2_test_fixture.collection(library)
+        work = odl2_test_fixture.work(collection)
+        license = odl2_test_fixture.license(work)
+        api = odl2_test_fixture.api(collection)
+        pool = odl2_test_fixture.pool(license)
+
+        data_source = DataSource.lookup(db.session, "Feedbooks", autocreate=True)
+        DatabaseTransactionFixture.set_settings(
+            collection.integration_configuration,
+            **{Collection.DATA_SOURCE_NAME_SETTING: data_source.name},
+        )
+        reaper = ODL2HoldReaper(db.session, collection, api=api)
+
+        now = utc_now()
+        yesterday = now - datetime.timedelta(days=1)
+
+        license.setup(concurrency=3, available=3)
+        expired_hold1, ignore = pool.on_hold_to(db.patron(), end=yesterday, position=0)
+        expired_hold2, ignore = pool.on_hold_to(db.patron(), end=yesterday, position=0)
+        expired_hold3, ignore = pool.on_hold_to(db.patron(), end=yesterday, position=0)
+        current_hold, ignore = pool.on_hold_to(db.patron(), position=3)
+        # This hold has an end date in the past, but its position is greater than 0
+        # so the end date is not reliable.
+        bad_end_date, ignore = pool.on_hold_to(db.patron(), end=yesterday, position=4)
+
+        progress = reaper.run_once(reaper.timestamp().to_data())
+
+        # The expired holds have been deleted and the other holds have been updated.
+        assert 2 == db.session.query(Hold).count()
+        assert [current_hold, bad_end_date] == db.session.query(Hold).order_by(
+            Hold.start
+        ).all()
+        assert 0 == current_hold.position
+        assert 0 == bad_end_date.position
+        assert current_hold.end > now
+        assert bad_end_date.end > now
+        assert 1 == pool.licenses_available
+        assert 2 == pool.licenses_reserved
+
+        # The TimestampData returned reflects what work was done.
+        assert "Holds deleted: 3. License pools updated: 1" == progress.achievements
+
+        # The TimestampData does not include any timing information --
+        # that will be applied by run().
+        assert None == progress.start
+        assert None == progress.finish
